requirements.txt
numpy
matplotlib
pysam
biopython